RAG/requirements.txt
streamlit>=1.31,<2
chromadb>=0.4.22,<0.6
langchain>=0.1.0,<0.3
langchain-community>=0.0.20,<0.3
sentence-transformers>=2.3,<3
openai>=1.12,<2
PyPDF2>=3,<4
python-docx>=1.1,<2
jieba>=0.42,<1
pydantic>=2.6,<3
python-dotenv>=1.0,<2
numpy>=1.26,<2
